scripts/requirements.txt
# Requirements for Campus Connect Backup/Restore Scripts
motor>=3.3.0
python-dotenv>=1.0.0
pymongo>=4.5.0